将单图文附件保存位置改为阿里云

+++ controllers/mp/matter/article.php
@@ -146,7 +146,7 @@ class resumable {
         }
     }
 }
-class resumableSae {
+class resumableAliOss {
     
     private $mpid;
     
@@ -181,7 +181,7 @@ class resumableSae {
      */
     private function createFileFromChunks($temp_dir, $articleid, $fileName, $chunkSize, $totalSize) 
     {
-        $fs = \TMS_APP::M('fs/attachment', $this->mpid);
+        $fs = \TMS_APP::M('fs/saestore', $this->mpid);
         // count all the parts of this file
         $total_files = 0;
         $rst = $fs->getListByPath($temp_dir);
@@ -193,15 +193,29 @@ class resumableSae {
         // check that all the parts are present
         // the size of the last part is between chunkSize and 2*$chunkSize
         if ($total_files * $chunkSize >=  ($totalSize - $chunkSize + 1)) {
-            // create the final destination file 
-            $dest = 'article_'.$articleid.'_'.$fileName;
+            
+            $fs2 = \TMS_APP::M('fs/alioss', $this->mpid, 'xxt-attachment');
+            $object = $this->mpid . '/article/' . $articleid . '/' . $_POST['resumableFilename'];
+            // uploadId
+            $upload = $fs2->initiate_multipart_upload($object);
+            if ($upload[0] === false) {
+                throw new \Exception($upload[1]);
+            }
+            $uploadId = $upload[1];
+            
+            $responseUploadPart = array();
+            // create the final destination file
             $content = '';
             for ($i=1; $i<=$total_files; $i++) {
                 $content .= $fs->read($temp_dir.'/'.$fileName.'.part'.$i);
                 $fs->delete($temp_dir.'/'.$fileName.'.part'.$i);
-                $this->_log('writing chunk '.$i);
             }
-            $fs->write($dest, $content);
+            $tmpfname = tempnam(sys_get_temp_dir(), 'xxt');
+            $handle = fopen($tmpfname, "w");
+            fwrite($handle, $content);
+            fclose($handle);
+            
+            $rsp = $fs2->create_mpu_object($object, $tmpfname);
         }
     }
     /**
@@ -231,7 +245,7 @@ class resumableSae {
             $temp_dir = $_POST['resumableIdentifier'];
             $dest_file = $temp_dir.'/'.$_POST['resumableFilename'].'.part'.$_POST['resumableChunkNumber'];
             // move the temporary file
-            $fs = \TMS_APP::M('fs/attachment', $this->mpid);
+            $fs = \TMS_APP::M('fs/saestore', $this->mpid);
             if (!$fs->upload($dest_file, $file['tmp_name'])) {
                 $this->_log('Error saving (move_uploaded_file) chunk '.$_POST['resumableChunkNumber'].' for file '.$_POST['resumableFilename']);
             } else {
@@ -579,7 +593,7 @@ class article extends matter_ctrl {
     public function upload_action($articleid)
     {
         if (defined('SAE_TMP_PATH'))
-            $resumable = new resumableSae($this->mpid);
+            $resumable = new resumableAliOss($this->mpid);
         else
             $resumable = new resumable($this->mpid);
             
@@ -587,7 +601,7 @@ class article extends matter_ctrl {
         exit;
     }
     /**
-     * 添加附件
+     * 上传成功后添加附件
      */
     public function attachmentAdd_action($id)
     {
@@ -595,7 +609,10 @@ class article extends matter_ctrl {
         /**
          * store to sae
          */
-        $url = 'article_'.$id.'_'.$file->name;
+        if (defined('SAE_TMP_PATH'))
+            $url = 'alioss://article/' . $id . '/' . $file->name;
+        else
+            $url = 'article_' . $id . '_' . $file->name;
         /**
          * store to local
          */
@@ -622,12 +639,18 @@ class article extends matter_ctrl {
      */
     public function attachmentDel_action($id)
     {
-        $att = $this->model()->query_obj_ss(array('url','xxt_article_attachment', "id='$id'"));
+        $att = $this->model()->query_obj_ss(array('name,url','xxt_article_attachment', "id='$id'"));
         /**
          * remove from fs
          */
-        $fs = $this->model('fs/attachment', $this->mpid);
-        $fs->delete($att->url);
+        if (strpos($att->url, 'alioss') === 0) {
+            $fs = $this->model('fs/alioss', $this->mpid, 'xxt-attachment');
+            $object = $this->mpid.'/article/'.$id.'/'.$att->name;
+            $fs->delete_object($object);
+        } else {
+            $fs = $this->model('fs/saestore', $this->mpid);
+            $fs->delete($att->url);
+        }
         /**
          * remove from local
          */
